Split up CRD structs to improve readability
